últimos cambios Stefany

+++ APIMARKET/Backend/Apimarket/Models/Feeding.Model.cs
@@ -20,10 +20,10 @@ namespace Apimarket.Model
         [DisplayFormat(DataFormatString = "{0:yyyy-MM-dd}", ApplyFormatInEditMode = true)]
 
         public DateTime Fec_Feeding { get; set; }
-        public int Can_Feeding { get; set; }
+        public decimal Can_Feeding { get; set; }
         //[DisplayName("Valor de la alimentación")]
 
-        public int Vlr_Feeding { get; set; }
+        public decimal Vlr_Feeding { get; set; }
         [DisplayName("Cantidad de alimentación")]
         //[Required(ErrorMessage = "El campo{0}es requerido")]
         //[StringLength(255, ErrorMessage = "El campo es {0} tiene un limite de caracteres de {1}")]
